Use single thread mode when no safe order_by can be used by multithreading
